requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
chardet==5.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.11
django-cors-headers==4.4.0
django-crontab==0.7.1
django-jazzmin==3.0.1
djangorestframework==3.15.2
drf-yasg==1.21.10
inflection==0.5.1
kombu==5.4.0
mysqlclient==2.2.7
numpy==2.1.2
packaging==25.0
pandas==2.2.3
pillow==11.0.0
prompt_toolkit==3.0.47
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2024.2
PyYAML==6.0.2
reportlab==4.2.5
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.14.1
tzdata==2024.1
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.13
